Created using spr 1.3.4

+++ bolt/include/bolt/Core/BinarySection.h
@@ -358,14 +358,14 @@ public:
   void clearRelocations();
 
   /// Add a new relocation at the given /p Offset.
-  void addRelocation(uint64_t Offset, MCSymbol *Symbol, uint64_t Type,
+  void addRelocation(uint64_t Offset, MCSymbol *Symbol, uint32_t Type,
                      uint64_t Addend, uint64_t Value = 0) {
     assert(Offset < getSize() && "offset not within section bounds");
     Relocations.emplace(Relocation{Offset, Symbol, Type, Addend, Value});
   }
 
   /// Add a dynamic relocation at the given /p Offset.
-  void addDynamicRelocation(uint64_t Offset, MCSymbol *Symbol, uint64_t Type,
+  void addDynamicRelocation(uint64_t Offset, MCSymbol *Symbol, uint32_t Type,
                             uint64_t Addend, uint64_t Value = 0) {
     addDynamicRelocation(Relocation{Offset, Symbol, Type, Addend, Value});
   }
